requirements.txt
click>=8.0
python-dotenv>=0.5.1
scikit-learn==0.24.1
pyyaml==3.11
marshmallow-dataclass==8.3.0
pandas==1.1.5
numpy==1.21.3
matplotlib==3.3.4
seaborn==0.11.1
